Add `\cvdoubleitemMy`: An extended version to allow modifying col. width
When `\hintscolumnwidth` was increased the `\cvdoubleitem` produced very
narrow columns for content as the hints width was taken to calculate the
remaining length.

But the first hint and the second does not sit in the same layout on the
page (they're both hints, but the second is not on the left border; so,
reusing the same length definition causes the narrow columns problem).